More import fixes after fixtures refactoring
